# txscript: Check equality via secp256k1 methods.